Generates ores in ChunkGenerator
Blocks cannot be set in populators so no ores were being generated. So I
moved ore generstion into the chunk generator.

WIP: Also, all entities (except Shulkers) are dying instantly because blocks are not being set to air in the entity populator. This needs fixing still.
